Added too 'printQuote' function to change background color with quote. Added comments, Added if statements to check if random quote has a "tag" property
I added a getRandomColor variable along with code calling the random color  to the printQuote function that allows the background color of the page to be changed as the quote refreshes or the button is pushed.

I also added a "tag" property to my array (for extra credit), meaning I also had to modify and add if statements that check for and print the appropriate information to the page.

To finish my project, I  added comments explaining my code and citing any snippets found on google.